+++ pyleco/test.py
@@ -22,11 +22,12 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 from typing import Any, Optional, Sequence, Union
 
 from .core.message import Message
 from .core.internal_protocols import CommunicatorProtocol
-from .core.rpc_generator import RPCGenerator
+from .json_utils.rpc_generator import RPCGenerator
 
 
 class FakeContext: