+++ tests/acceptance_tests/test_proxy_server_live.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 import pytest
 from time import sleep