+++ tests/utils/test_base_communicator.py
@@ -22,6 +22,7 @@
 # THE SOFTWARE.
 #
 
+from __future__ import annotations
 import logging
 import time
 from typing import Optional
@@ -31,7 +32,7 @@ import pytest
 
 from pyleco.test import FakeSocket
 from pyleco.core.message import Message, MessageTypes
-from pyleco.core.rpc_generator import RPCGenerator
+from pyleco.json_utils.rpc_generator import RPCGenerator
 from pyleco.errors import DUPLICATE_NAME, ErrorResponse
 
 from pyleco.utils.base_communicator import BaseCommunicator