+++ tests/utils/test_communicator.py
@@ -24,13 +24,13 @@
 
 from unittest.mock import MagicMock
 
-from jsonrpcobjects.errors import JSONRPCError
 import pytest
 
 from pyleco.core import VERSION_B
 from pyleco.core.message import Message, MessageTypes
 from pyleco.errors import NOT_SIGNED_IN, NODE_UNKNOWN
 from pyleco.core.serialization import serialize_data
+from pyleco.json_utils.errors import JSONRPCError
 
 from pyleco.utils.communicator import Communicator
 from pyleco.test import FakeSocket, FakeContext